Update meeting contacts as trivial change to republish
Trivial change to trigger a republish of the charm docs.

http://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html

Note that on July 3rd of 2017, all projects have been moved
to a new location on this server.

Each project is required to merge any single change in order
to be published. Keep in mind that bookmarks saved before the
changes may no longer function. See the specification for
more information.

Change-Id: I9816f5c5e81b8fab10edc328f2c0bda94fcac7a9