Add relnote for ncc around setting cache-known-hosts to true
We are "flipping the switch" on the default for cache-known-hosts
to true for 19.10 (was false for 19.07) and so this release note
indicates this change, particularly around upgrade and the change in
behaviour.

See I644df508563f93b661af6bfd5fb668f0e31d5e4d for the
nova-cloud-controller charm change.

Change-Id: Ib3d17b4057dd7f671f123b1da7df21c9f3d7e182