Polish 1910 release notes - part I
This doesn't add anything qualitatively different. It
adds polish in terms of formatting, slight word changes,
fixes broken links, and moves the order of a paragraph.

The Upgrading OpenStack section re placement was changed
to sync to what is in app-upgrade-openstack.rst (CDG).

Removed text on Ceph was transferred to
app-upgrade-openstack.rst.

Change-Id: I9c6985ec1dfa6b10485c6b46299aa41f36a332c0