Refresh intro pages
Remove deployment page and merge contents with the
getting-started page.

Enhance the charm-store page and rename it to the
charms-and-bundles page. When ubuntu.com/openstack/docs
is ready the contents of this page will be moved to it
(and linked from getting-started).

Remove mailing-list page and merge contents with the
find-us page.

Remove meetings page as per team consensus.

Changes necessary on a few pages to compensate for
the above two changes/lines.

A few minor drive-by spelling and format fixes.

Closes-Bug: #1921371
Change-Id: Icad0c415f324e818cb18d5abe61b475ce903ca11