Add notice regarding Xenial gate
Add a deprecation notice indicating the removal of regular gate
testing for the Xenial series in the Yoga timeline. When a migration
to charmhub occurs, a separate series of tracks can be created for the
Xenial related OpenStack releases with specific gates for those.

The notice also indicates that the nature of maintenance will be limited
to critical and high security fixes only.

Change-Id: I395b01b1d321b198aa551f2e01fb0ff02da2bdd2