Add intersphinx extension
Add the Sphinx extension 'intersphinx' and apply it to
the CDG so that linking to the latter can be more seamless.

Test out intersphinx features on the release-notes-template
page.

The remote intersphinx "inventory objects file" objects.inv
has to be hardcoded to the 'latest' release during the
development cyle. This should be updated to the latest
OpenStack release when a new release is cut for the remote
guide.

Fix a few broken links on the release-notes-template page.

Change-Id: I021697c394b83d2448dc21add945ab8fd9d39a3f