Add new charm delivery page - Charmhub
This is a new project page. It gives a high level overview of
how the OpenStack charms will be delivered using the Charmhub
from the end-user's perspective. A developer-centric page may
be included at a later time.

The scafolding for the 'About the project' needed to be added
as this is the first page under this category. Most current
"reference" pages are mis-categorised and will eventually be
moved into it.

The tabs feature was added to Sphinx in order to provide a
reasonable display of the charm channel information.

The `latest` track was greatly de-emphasised due to its
inherent hazardous nature.

There is a reference to "a special charm operation" that is
currently unpublished, but that is where details on migration
will be stored (currently in the CDG).

Change-Id: Ia3ebfad4a6aa43620cad188b07ec6fc7a29b71a7