Rewrite getting started
Rewrite the getting started pages by turning it
into an actual tutorial. This will be the main
Charmed OpenStack tutorial (golden path) for
people to follow.

It is hoped that once bundles in the Charmhub are
properly staged the need to download the bundle
repository will no longer be necessary.

Change-Id: I099a58874fc963688ae18d7353aa8d086de98da4