Copy issues and procedures from deploy-guide
This PR is part of the ongoing work to migrate pages
from the deploy-guide to the charm-guide that are not
specific to a manual test cloud deployment.

Some links necessarily needed to be changed in some
files.

A link was fixed on the placement-charm page
(reference to paused-single-unit in the deploy-guide).

Change-Id: Iba9231ba501253692acba9f059a913b54226cb5e