Review and refactor upgrade pages
Review and refactor upgrade pages to make it clearer what
values to use for the combination of charm channels, series,
and OpenStack release.

In order for this to work well (avoid duplication) for the
various upgrade types as well as for links from the
deploy-guide, central, and purely informative, content was
moved into two new conceptual pages. In so doing, the lengthy
upgrade pages were made lighter.

A bug was incidentally closed.

Closes-Bug: #1974255
Change-Id: I6a493b4dcc8c11a11a41bcafd91596ef0ce80122