Add linkchecking
Enable the built-in Sphinx 'linkcheck' builder and
fix initial findings.

Add this as an opportunistic build target only. The process is
long, and false positives are common (mostly anchor-related
for some reason). There is also already a third-party
company-wide service in place. This builder however will lead
to better link hygiene as it is stricter (mostly for anchors).
It also records all redirects, which should eventually be
replaced by proper links.

Change-Id: Ibed74b9f820c17a5984a78f4f61153e0499453fa