Add zed-jammy to charm delivery page
Add zed track for jammy series on charm delivery page.

Note that pacemaker-remote does not appear to have a
stable revision on either focal or jammy.

We should probably not continue to state mulitple upstream
versions just because they work but stick with a
recommended version (see OVN and Vault).

Change-Id: I3d4c5db75548a44072aa864f0bc3c97e4601997d