Add most recent revisions of 21.10 charms
This is to provide a permanent record of the 21.10 charms, and prior
releases, against series for each of the charms from the latest/stable
channel.

Although the 'latest' track will be closed and this document refers to
that, other pages impacted by this event will be addressed in a separate
change.

Change-Id: Id31eb86843b04cc778297a8597495c92650fe714