Update the backport policy core voting requirements
This patch updates the core voting requirements so that library
backports only require a single review from a core reviewer/stable
maintainer to land the review on a stable branch.  This actually makes
it the same as the master/main development branch for most libraries as
in only a single review is required.

Change-Id: I250ceac19865e0daa1e4aaa32d1d58b129656776