Remove wait option from all run commands
The Charm Guide was recently updated for Juju 3.x. However, it was not
known at the time that the semantics of the --wait option for the `juju
run` command had changed.

In 3, the run command by default stays in the foreground (a --background
option has been added to achieve previous 2.9 default behaviour). The
--wait option is now a timeout and, if used, requires a value. Since
the previous PR simply substituted commands, all `juju run --wait`
commands will fail.

Change-Id: I6bb90762ad5cb5ca97ca311501b1ff7d3d9a3ccb
Signed-off-by: Peter Matulis <peter.matulis@canonical.com>